+++ modbusTcpHelper/src/main/java/com/zhipuchina/function/ServerFunctionWriteTemplate.java
@@ -2,8 +2,8 @@ package com.zhipuchina.function;
 
 import com.zhipuchina.exception.ModbusException;
 import com.zhipuchina.handler.ModbusTcpBasicSession;
-import com.zhipuchina.model.MemoryTypes;
 import com.zhipuchina.model.Buffer;
+import com.zhipuchina.model.MemoryTypes;
 import com.zhipuchina.utils.ConvertTo;
 
 public abstract class ServerFunctionWriteTemplate implements FunctionController{
@@ -14,14 +14,14 @@ public abstract class ServerFunctionWriteTemplate implements FunctionController{
         this.type = type;
     }
 
-    public abstract byte[] getValue(byte[] ADU);
+    public abstract Integer[] getValue(byte[] ADU);
     public abstract int getCount(byte[] ADU);
 
     @Override
     public byte[] serve(byte[] header, byte[] ADU, ModbusTcpBasicSession session) throws ModbusException {
         //地址
         int address = ConvertTo.getShort(ADU[1],ADU[2]);
-        byte[] value = getValue(ADU);
+        Integer[] value = getValue(ADU);
         //写入
         Buffer.setValue(type,address,getCount(ADU),value);
         byte[] out = new byte[12];
